Separate integration tests from unit tests
The idea here is to colocate the unit tests with the code under tests,
which is generally what Facebook crew is reccomending. The integration
test, which test more than one piece of code then live in the the `test`
folder.

Integration tests can now be run with `yarn integration` while unit
tests can be run with `yarn test` as normal.
