+++ src/schema/types/__tests__/Latitude.test.js
@@ -1,4 +1,4 @@
-import Latitude from '../src/schema/types/Latitude'
+import Latitude from '../Latitude'
 import {
   graphql,
   GraphQLSchema,
